Automatically restart prometheus in Makefile
Whenever prometheus's config map changes it needs to be reloaded for it to take effect.

Just deleting the pod works for now, but unfortunately this also deletes the locally stored metrics history. This data is pulled into Azure Monitor anyway so we wouldn't lose it completely but it's still not ideal.

Prometheus does have a `/-/reload` HTTP endpoint for reloading the config but it's disabled by default, and I haven't found if/how to enable it through the istio profile yet.

https://prometheus.io/docs/prometheus/latest/configuration/configuration/